Move Adapter classes into subpackage.
