+++ source/src/in/joind/EventComments.java
@@ -27,8 +27,10 @@ import org.json.JSONObject;
 
 import java.util.ArrayList;
 
+import in.joind.adapter.EventCommentAdapter;
+
 public class EventComments extends JIActivity implements OnClickListener {
-    private JIEventCommentAdapter m_eventCommentAdapter;    // adapter for listview
+    private EventCommentAdapter m_eventCommentAdapter;    // adapter for listview
     private JSONObject eventJSON;
     private int event_id;
 
@@ -56,7 +58,7 @@ public class EventComments extends JIActivity implements OnClickListener {
 
         // Initialize comment list
         ArrayList<JSONObject> m_eventcomments = new ArrayList<>();
-        m_eventCommentAdapter = new JIEventCommentAdapter(this, R.layout.talkrow, m_eventcomments);
+        m_eventCommentAdapter = new EventCommentAdapter(this, R.layout.talkrow, m_eventcomments);
         final PullToRefreshListView eventcommentlist = (PullToRefreshListView) findViewById(R.id.EventDetailComments);
         eventcommentlist.setAdapter(m_eventCommentAdapter);
 
@@ -214,56 +216,4 @@ public class EventComments extends JIActivity implements OnClickListener {
         }.start();
     }
 
-}
-
-
-/**
- * Adapter that hold our event comment rows. See  JIEventAdapter class in main.java for more info
- */
-class JIEventCommentAdapter extends ArrayAdapter<JSONObject> {
-    private ArrayList<JSONObject> items;
-    private Context context;
-    private ImageLoader image_loader;            // gravatar image loader
-
-    public JIEventCommentAdapter(Context context, int textViewResourceId, ArrayList<JSONObject> items) {
-        super(context, textViewResourceId, items);
-        this.context = context;
-        this.items = items;
-
-        this.image_loader = new ImageLoader(context.getApplicationContext(), "gravatars");
-    }
-
-    public View getView(int position, View convertview, ViewGroup parent) {
-        View v = convertview;
-        if (v == null) {
-            LayoutInflater vi = (LayoutInflater) this.context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            v = vi.inflate(R.layout.commentrow, parent, false);
-        }
-
-        JSONObject o = items.get(position);
-        if (o == null) return v;
-
-        ImageView el = (ImageView) v.findViewById(R.id.CommentRowGravatar);
-        el.setTag("");
-        el.setVisibility(View.GONE);
-
-        if (o.optInt("user_id") > 0) {
-            String filename = "user" + o.optString("user_id") + ".jpg";
-            el.setTag(filename);
-            image_loader.displayImage("http://joind.in/inc/img/user_gravatar/", filename, (Activity) context, el);
-        }
-
-        String commentDate = DateHelper.parseAndFormat(o.optString("created_date"), "d LLL yyyy");
-        TextView t1 = (TextView) v.findViewById(R.id.CommentRowComment);
-        TextView t2 = (TextView) v.findViewById(R.id.CommentRowUName);
-        TextView t3 = (TextView) v.findViewById(R.id.CommentRowDate);
-        if (t1 != null) t1.setText(o.optString("comment"));
-        if (t2 != null) t2.setText(o.isNull("user_display_name") ? "(" + this.context.getString(R.string.generalAnonymous) + ") " : o.optString("user_display_name") + " ");
-        if (t3 != null) t3.setText(commentDate);
-
-        ImageView r = (ImageView) v.findViewById(R.id.CommentRowRate);
-        r.setVisibility(View.GONE);
-
-        return v;
-    }
 }
